% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dh_cums.R
\name{dh_join_cums}
\alias{dh_join_cums}
\title{join the normal and observed daily rainfall cumuls}
\usage{
dh_join_cums(obs, norm)
}
\arguments{
\item{obs}{dataframe}

\item{norm}{dataframe}
}
\value{
dataframe
}
