% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhw_map.R
\name{dhw_idw}
\alias{dhw_idw}
\title{spatialize the Pameseb deficit hydrique with IDW algo}
\usage{
dhw_idw(data, param, grid)
}
\arguments{
\item{data}{sp dataframe}

\item{param}{cahracter}
}
\value{
sp dataframe
}
