% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhw_cums.R
\name{dhw_join_cums}
\alias{dhw_join_cums}
\title{join the normal and observed daily rainfall cumuls}
\usage{
dhw_join_cums(obs, norm)
}
\arguments{
\item{obs}{dataframe}

\item{norm}{dataframe}
}
\value{
dataframe
}
