% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhw_map.R
\name{dhw_map}
\alias{dhw_map}
\title{make the static map of the deficit hydrique}
\usage{
dhw_map(gridded.data.df, boundaries.sf, title.chr, legend.chr, target.chr,
  nb_classes.num, reverse_pal.bool, pretty_breaks.bool)
}
\arguments{
\item{gridded.data.df}{dataframe}

\item{title.chr}{chr}

\item{legend.chr}{chr}

\item{target.chr}{chr}

\item{nb_classes.num}{num}

\item{reverse_pal.bool}{bool}
}
\value{
dataframe
}
