% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhw_plot.R
\name{dhw_plot}
\alias{dhw_plot}
\title{plot the deficit hydrique from Ernage data for the period of interest}
\usage{
dhw_plot(data, int)
}
\arguments{
\item{data}{dataframe}

\item{int}{boolean}
}
\value{
ggplot graph
}
