% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhw_filter.R
\name{dhw_filter}
\alias{dhw_filter}
\title{filter the observed data for the period of interest and relevant stations}
\usage{
dhw_filter(data, dfrom, dto)
}
\arguments{
\item{data}{dataframe}

\item{dfrom}{"YYYY-MM-DD"}

\item{dto}{"YYYY-MM-DD"}
}
\value{
dataframe
}
